# MQE: don't create copies of native histogram during binary operations