--- a/pkg/streamingpromql/operators/binops/binary_operation.go
+++ b/pkg/streamingpromql/operators/binops/binary_operation.go
@@ -512,7 +512,7 @@ var arithmeticAndComparisonOperationFuncs = map[parser.ItemType]binaryOperationF
 
 		return 0, nil, false, false, nil
 	},
-	parser.DIV: func(lF, rF float64, lH, rH *histogram.FloatHistogram, canMutateLeft, canMutateRight bool) (float64, *histogram.FloatHistogram, bool, bool, error) {
+	parser.DIV: func(lF, rF float64, lH, rH *histogram.FloatHistogram, canMutateLeft, _ bool) (float64, *histogram.FloatHistogram, bool, bool, error) {
 		if lH == nil && rH == nil {
 			return lF / rF, nil, true, true, nil
 		}
@@ -527,21 +527,21 @@ var arithmeticAndComparisonOperationFuncs = map[parser.ItemType]binaryOperationF
 
 		return 0, nil, false, false, nil
 	},
-	parser.POW: func(lF, rF float64, lH, rH *histogram.FloatHistogram, canMutateLeft, canMutateRight bool) (float64, *histogram.FloatHistogram, bool, bool, error) {
+	parser.POW: func(lF, rF float64, lH, rH *histogram.FloatHistogram, _, _ bool) (float64, *histogram.FloatHistogram, bool, bool, error) {
 		if lH == nil && rH == nil {
 			return math.Pow(lF, rF), nil, true, true, nil
 		}
 
 		return 0, nil, false, false, nil
 	},
-	parser.MOD: func(lF, rF float64, lH, rH *histogram.FloatHistogram, canMutateLeft, canMutateRight bool) (float64, *histogram.FloatHistogram, bool, bool, error) {
+	parser.MOD: func(lF, rF float64, lH, rH *histogram.FloatHistogram, _, _ bool) (float64, *histogram.FloatHistogram, bool, bool, error) {
 		if lH == nil && rH == nil {
 			return math.Mod(lF, rF), nil, true, true, nil
 		}
 
 		return 0, nil, false, false, nil
 	},
-	parser.ATAN2: func(lF, rF float64, lH, rH *histogram.FloatHistogram, canMutateLeft, canMutateRight bool) (float64, *histogram.FloatHistogram, bool, bool, error) {
+	parser.ATAN2: func(lF, rF float64, lH, rH *histogram.FloatHistogram, _, _ bool) (float64, *histogram.FloatHistogram, bool, bool, error) {
 		if lH == nil && rH == nil {
 			return math.Atan2(lF, rF), nil, true, true, nil
 		}
@@ -559,14 +559,14 @@ var arithmeticAndComparisonOperationFuncs = map[parser.ItemType]binaryOperationF
 
 		if lH != nil && rH != nil {
 			if lH.Equals(rH) {
-				res := lH
+				var res *histogram.FloatHistogram
 
-				if !canMutateLeft {
-					if canMutateRight {
-						res = rH
-					} else {
-						res = lH.Copy()
-					}
+				if canMutateLeft {
+					res = lH
+				} else if canMutateRight {
+					res = rH
+				} else {
+					res = lH.Copy()
 				}
 
 				return 0, res, true, true, nil
@@ -577,7 +577,7 @@ var arithmeticAndComparisonOperationFuncs = map[parser.ItemType]binaryOperationF
 
 		return 0, nil, false, false, nil
 	},
-	parser.NEQ: func(lF, rF float64, lH, rH *histogram.FloatHistogram, canMutateLeft, canMutateRight bool) (float64, *histogram.FloatHistogram, bool, bool, error) {
+	parser.NEQ: func(lF, rF float64, lH, rH *histogram.FloatHistogram, canMutateLeft, _ bool) (float64, *histogram.FloatHistogram, bool, bool, error) {
 		if lH == nil && rH == nil {
 			if lF != rF {
 				return lF, nil, true, true, nil
@@ -600,7 +600,7 @@ var arithmeticAndComparisonOperationFuncs = map[parser.ItemType]binaryOperationF
 
 		return lF, lH, false, false, nil
 	},
-	parser.LTE: func(lF, rF float64, lH, rH *histogram.FloatHistogram, canMutateLeft, canMutateRight bool) (float64, *histogram.FloatHistogram, bool, bool, error) {
+	parser.LTE: func(lF, rF float64, lH, rH *histogram.FloatHistogram, _, _ bool) (float64, *histogram.FloatHistogram, bool, bool, error) {
 		if lH != nil || rH != nil {
 			return 0, nil, false, false, nil
 		}
@@ -611,7 +611,7 @@ var arithmeticAndComparisonOperationFuncs = map[parser.ItemType]binaryOperationF
 
 		return 0, nil, false, true, nil
 	},
-	parser.LSS: func(lF, rF float64, lH, rH *histogram.FloatHistogram, canMutateLeft, canMutateRight bool) (float64, *histogram.FloatHistogram, bool, bool, error) {
+	parser.LSS: func(lF, rF float64, lH, rH *histogram.FloatHistogram, _, _ bool) (float64, *histogram.FloatHistogram, bool, bool, error) {
 		if lH != nil || rH != nil {
 			return 0, nil, false, false, nil
 		}
@@ -622,7 +622,7 @@ var arithmeticAndComparisonOperationFuncs = map[parser.ItemType]binaryOperationF
 
 		return 0, nil, false, true, nil
 	},
-	parser.GTE: func(lF, rF float64, lH, rH *histogram.FloatHistogram, canMutateLeft, canMutateRight bool) (float64, *histogram.FloatHistogram, bool, bool, error) {
+	parser.GTE: func(lF, rF float64, lH, rH *histogram.FloatHistogram, _, _ bool) (float64, *histogram.FloatHistogram, bool, bool, error) {
 		if lH != nil || rH != nil {
 			return 0, nil, false, false, nil
 		}
@@ -633,7 +633,7 @@ var arithmeticAndComparisonOperationFuncs = map[parser.ItemType]binaryOperationF
 
 		return 0, nil, false, true, nil
 	},
-	parser.GTR: func(lF, rF float64, lH, rH *histogram.FloatHistogram, canMutateLeft, canMutateRight bool) (float64, *histogram.FloatHistogram, bool, bool, error) {
+	parser.GTR: func(lF, rF float64, lH, rH *histogram.FloatHistogram, _, _ bool) (float64, *histogram.FloatHistogram, bool, bool, error) {
 		if lH != nil || rH != nil {
 			return 0, nil, false, false, nil
 		}
@@ -647,7 +647,7 @@ var arithmeticAndComparisonOperationFuncs = map[parser.ItemType]binaryOperationF
 }
 
 var boolComparisonOperationFuncs = map[parser.ItemType]binaryOperationFunc{
-	parser.EQLC: func(lF, rF float64, lH, rH *histogram.FloatHistogram, canMutateLeft, canMutateRight bool) (float64, *histogram.FloatHistogram, bool, bool, error) {
+	parser.EQLC: func(lF, rF float64, lH, rH *histogram.FloatHistogram, _, _ bool) (float64, *histogram.FloatHistogram, bool, bool, error) {
 		if lH == nil && rH == nil {
 			if lF == rF {
 				return 1, nil, true, true, nil
@@ -666,7 +666,7 @@ var boolComparisonOperationFuncs = map[parser.ItemType]binaryOperationFunc{
 
 		return 0, nil, false, false, nil
 	},
-	parser.NEQ: func(lF, rF float64, lH, rH *histogram.FloatHistogram, canMutateLeft, canMutateRight bool) (float64, *histogram.FloatHistogram, bool, bool, error) {
+	parser.NEQ: func(lF, rF float64, lH, rH *histogram.FloatHistogram, _, _ bool) (float64, *histogram.FloatHistogram, bool, bool, error) {
 		if lH == nil && rH == nil {
 			if lF != rF {
 				return 1, nil, true, true, nil
@@ -685,7 +685,7 @@ var boolComparisonOperationFuncs = map[parser.ItemType]binaryOperationFunc{
 
 		return 0, nil, false, false, nil
 	},
-	parser.LTE: func(lF, rF float64, lH, rH *histogram.FloatHistogram, canMutateLeft, canMutateRight bool) (float64, *histogram.FloatHistogram, bool, bool, error) {
+	parser.LTE: func(lF, rF float64, lH, rH *histogram.FloatHistogram, _, _ bool) (float64, *histogram.FloatHistogram, bool, bool, error) {
 		if lH != nil || rH != nil {
 			return 0, nil, false, false, nil
 		}
@@ -696,7 +696,7 @@ var boolComparisonOperationFuncs = map[parser.ItemType]binaryOperationFunc{
 
 		return 0, nil, true, true, nil
 	},
-	parser.LSS: func(lF, rF float64, lH, rH *histogram.FloatHistogram, canMutateLeft, canMutateRight bool) (float64, *histogram.FloatHistogram, bool, bool, error) {
+	parser.LSS: func(lF, rF float64, lH, rH *histogram.FloatHistogram, _, _ bool) (float64, *histogram.FloatHistogram, bool, bool, error) {
 		if lH != nil || rH != nil {
 			return 0, nil, false, false, nil
 		}
@@ -707,7 +707,7 @@ var boolComparisonOperationFuncs = map[parser.ItemType]binaryOperationFunc{
 
 		return 0, nil, true, true, nil
 	},
-	parser.GTE: func(lF, rF float64, lH, rH *histogram.FloatHistogram, canMutateLeft, canMutateRight bool) (float64, *histogram.FloatHistogram, bool, bool, error) {
+	parser.GTE: func(lF, rF float64, lH, rH *histogram.FloatHistogram, _, _ bool) (float64, *histogram.FloatHistogram, bool, bool, error) {
 		if lH != nil || rH != nil {
 			return 0, nil, false, false, nil
 		}
@@ -718,7 +718,7 @@ var boolComparisonOperationFuncs = map[parser.ItemType]binaryOperationFunc{
 
 		return 0, nil, true, true, nil
 	},
-	parser.GTR: func(lF, rF float64, lH, rH *histogram.FloatHistogram, canMutateLeft, canMutateRight bool) (float64, *histogram.FloatHistogram, bool, bool, error) {
+	parser.GTR: func(lF, rF float64, lH, rH *histogram.FloatHistogram, _, _ bool) (float64, *histogram.FloatHistogram, bool, bool, error) {
 		if lH != nil || rH != nil {
 			return 0, nil, false, false, nil
 		}