@charleskorn 2024-12-13
This is ready for a review, but will require changes after https://github.com/grafana/mimir/pull/10119 is merged.