@charleskorn 2025-01-20
> I haven't looked closely, but do we have good tests to that would fail if we weren't nil'ing values before returning to the pool appropriately? (Just thinking how we can better test this).

Yep, the mangling of native histograms when slices of `HPoint` are returned to the pool catches any cases of this. (The tests I added in this PR were all failing before I added the `nil`-ing, as well as the equivalent cases for one-to-one matching and vector/scalar cases as well.)

> I also imagine this would not work if we every revisited trying not to copy histograms on lookbacks. It might be worth adding a comment somewhere as such.

I'm not sure we'd ever go back to this, and a bunch of stuff will now break if we ever do this - the mixed metrics gauntlet you added should catch these, but if we ever revisit this we're going to have to check a lot of different places carefully. So I'm not sure it's worth adding a comment.